+++ R/0_imports.R
@@ -19,13 +19,13 @@ if (getRversion() >= "2.15.1") {
 #' @importFrom furrr future_map future_pmap furrr_options
 #' @importFrom progressr with_progress
 #' @importFrom stringr str_count str_pad str_length
-#' @importFrom rlang := .data enexpr env_has
+#' @importFrom rlang := .data enexpr env_has .env
 #' @importFrom dplyr select mutate filter summarise group_by summarise_all summarise_at
 #' @importFrom dplyr mutate_all mutate_at ungroup vars bind_cols bind_rows pull
 #' @importFrom dplyr n_distinct distinct arrange right_join left_join inner_join
 #' @importFrom dplyr rename sample_n between row_number dense_rank desc case_when
 #' @importFrom dplyr group_by_at n is_grouped_df everything one_of lag any_of across
-#' @importFrom dplyr lead all_equal
+#' @importFrom dplyr lead
 #' @importFrom tidyr tibble unnest nest fill
 #' @importFrom knitr kable
 #' @importFrom tibble tibble