+++ NEWS.md
@@ -1,5 +1,5 @@
 # simaerep 0.4.3
-- delete performance unit tests (poisson faster than bootstrap) to accomadate CRAN request
+- delete performance unit tests (poisson faster than bootstrap) to accommodate CRAN request
 
 # simaerep 0.4.2
 - CRAN submission